+++ packages/xchain-bitcoincash/src/client.ts
@@ -404,10 +404,12 @@ class Client implements BitcoinCashClient, XChainClient {
    * @param {TxParams&FeeRate} params The transfer options.
    * @returns {TxHash} The transaction hash.
    */
-  transfer = async (params: TxParams & { feeRate: FeeRate }): Promise<TxHash> => {
+  transfer = async (params: TxParams & { feeRate?: FeeRate }): Promise<TxHash> => {
     try {
+      const feeRate = params.feeRate || (await this.getFeeRates()).fast
       const tx = await utils.buildTx({
         ...params,
+        feeRate,
         sender: this.getAddress(),
         clientUrl: this.getClientURL(),
         network: this.network,