+++ packages/xchain-litecoin/src/client.ts
@@ -397,10 +397,12 @@ class Client implements LitecoinClient, XChainClient {
    * @param {TxParams&FeeRate} params The transfer options.
    * @returns {TxHash} The transaction hash.
    */
-  transfer = async (params: TxParams & { feeRate: FeeRate }): Promise<TxHash> => {
+  transfer = async (params: TxParams & { feeRate?: FeeRate }): Promise<TxHash> => {
     try {
+      const feeRate = params.feeRate || (await this.getFeeRates()).fast
       const { psbt } = await Utils.buildTx({
         ...params,
+        feeRate,
         sender: this.getAddress(),
         nodeUrl: this.nodeUrl,
         network: this.net,